feat(portal): inline AgreementSection (email-matched signer token via overview)
